update cấu trúc dữ liệu gửi đi

+++ Client/src/domain/Packages.java
@@ -5,11 +5,6 @@
  */
 package domain;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.io.Serializable;
 
 /**
@@ -47,7 +42,7 @@ public class Packages<T> implements Serializable {
         return status;
     }
 
-    public void setStatus(short status) {
+    public void setStatus(int status) {
         this.status = status;
     }
